Once again, poor commit discipline
Very much still in the "Run really fast and break everything" stage. I'm
not really able to do Just One Thing yet, so for the time being commit
discipline will be poor.

In this commit we have the beginnings of an AngularJS app, a dramatic
overhaul of Charts and their implementation, and the stabilization of
the 'GET' endpoints for /charts/4-2 and /charts/4-3. 'POST' for those
endpoints remains entirely unimplemented.